[] DocDB: Fixes infinite loop detected error
Summary:
This diff fixes the infinite loop detected error which occurs when system keys are encountered in a reverse scan. The bug was that in reverse scan, instead of seeking to the previous DocKey, it was seeking
to the next one causing a infinite loop.

Test Plan: ./yb_build.sh --skip-java --cxx-test docrowwiseiterator-test --gtest-filter=DocRowwiseIteratorTest.ClusteredFilterRangeWithTableTombstoneReverseScan

Differential Revision: https://phabricator.dev.yugabyte.com/D22098