[MERGE PG15] test: fix python doctest failure
Summary:
build-support/jenkins/build-and-test.sh involves running doctest on
select python files.  This is what jenkins runs, whereas developers
typically run ./yb_build.sh.

doctest appears to have a strange bug where it does not like it when
sys.stdout is set to something else by the time of the end of file.
This doesn't seem to matter when executing the python script itself: it
only matters for doctest.  PG 15 merge involves updating
generate_unaccent_rules.py, and this update hits the above bug.  Appease
doctest by restoring sys.stdout to the original object at the end of the
file.

It seems wasteful to run doctest on files that don't have docstrings,
but that can be fixed in the future.

Test Plan: jenkins: compile only, rebase: pg15

Reviewers: mbautin, neil, steve.varnau

Reviewed By: steve.varnau

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D23266