[pg15] fix: Activate function yb_ipath_matches_pk
Summary:
The initial merge commit disabled the function yb_ipath_matches_pk due to compilation issue(s). The issue was caused due to changes in PG commit 1a8d5afb0dfc5d0dcc6eda0656a34cb1f0cf0bdf (titled, "Refactor the representation of indexable clauses in IndexPaths").

Fix the function as per the upstream change and re-enable it. With this, a single RPC is used for locking rows in non-serializable isolation level when PK is specified and yb_lock_pk_single_rpc is true (introduced in D26431).

As a result of the change, test TestConflictRetriesOnDocdb passes, add it to the passing test list.

A similar change was done in the function yb_single_row_update_or_delete_path in D27692. Modify it further to:
- add an assertion that indexcols is NULL
- minor refactor: fix a line limit violation, remove two unnecessary variable declarations

The function yb_cost_index requires a similar change, which will be done separately.

Test Plan:
Jenkins: rebase: pg15
./pg15_tests/test_lint.sh
./yb_build.sh --cxx-test pgwrapper_pg_read_time-test --gtest_filter PgReadTimeTest.TestConflictRetriesOnDocdb -n 10

Reviewers: tfoucher, jason

Reviewed By: tfoucher

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D32862